impr: Re-render SVG textures on scaling changes
This change also future-proofs the scaling event so it can be fired
when users change the scaling mode in the settings, since both user
scaling and window content scaling changes require
re-rasterisation.

+++ lib/libimhex/source/ui/imgui_imhex_extensions.cpp
@@ -33,7 +33,7 @@ namespace ImGuiExt {
 
     namespace {
 
-        void adjustSVGScale(const lunasvg::Document *document, int &width, int &height, int scale) {
+        void adjustSVGScale(const lunasvg::Document *document, int &width, int &height, float scale) {
             if (document->width() == 0 || document->height() == 0)
                 return;
